+++ src/models/components/queryconversationsresponsedata.ts
@@ -144,10 +144,10 @@ export type ConversationsResultItem = {
   maxPropertiesSupportAgentName?: number | undefined;
   maxUserPropertiesUserId?: number | undefined;
   maxUserPropertiesSupportAgentName?: number | undefined;
-  propertiesUserId?: string | undefined;
-  propertiesSupportAgentName?: string | undefined;
-  userPropertiesUserId?: string | undefined;
-  userPropertiesSupportAgentName?: string | undefined;
+  propertiesUserId?: string | null | undefined;
+  propertiesSupportAgentName?: string | null | undefined;
+  userPropertiesUserId?: string | null | undefined;
+  userPropertiesSupportAgentName?: string | null | undefined;
 };
 
 /**
@@ -318,10 +318,10 @@ export const ConversationsResultItem$inboundSchema: z.ZodType<
   "max_properties.supportAgentName": z.number().optional(),
   "max_userProperties.userId": z.number().optional(),
   "max_userProperties.supportAgentName": z.number().optional(),
-  "properties.userId": z.string().optional(),
-  "properties.supportAgentName": z.string().optional(),
-  "userProperties.userId": z.string().optional(),
-  "userProperties.supportAgentName": z.string().optional(),
+  "properties.userId": z.nullable(z.string()).optional(),
+  "properties.supportAgentName": z.nullable(z.string()).optional(),
+  "userProperties.userId": z.nullable(z.string()).optional(),
+  "userProperties.supportAgentName": z.nullable(z.string()).optional(),
 }).transform((v) => {
   return remap$(v, {
     "id_hour": "idHour",
@@ -560,10 +560,10 @@ export type ConversationsResultItem$Outbound = {
   "max_properties.supportAgentName"?: number | undefined;
   "max_userProperties.userId"?: number | undefined;
   "max_userProperties.supportAgentName"?: number | undefined;
-  "properties.userId"?: string | undefined;
-  "properties.supportAgentName"?: string | undefined;
-  "userProperties.userId"?: string | undefined;
-  "userProperties.supportAgentName"?: string | undefined;
+  "properties.userId"?: string | null | undefined;
+  "properties.supportAgentName"?: string | null | undefined;
+  "userProperties.userId"?: string | null | undefined;
+  "userProperties.supportAgentName"?: string | null | undefined;
 };
 
 /** @internal */
@@ -692,10 +692,10 @@ export const ConversationsResultItem$outboundSchema: z.ZodType<
   maxPropertiesSupportAgentName: z.number().optional(),
   maxUserPropertiesUserId: z.number().optional(),
   maxUserPropertiesSupportAgentName: z.number().optional(),
-  propertiesUserId: z.string().optional(),
-  propertiesSupportAgentName: z.string().optional(),
-  userPropertiesUserId: z.string().optional(),
-  userPropertiesSupportAgentName: z.string().optional(),
+  propertiesUserId: z.nullable(z.string()).optional(),
+  propertiesSupportAgentName: z.nullable(z.string()).optional(),
+  userPropertiesUserId: z.nullable(z.string()).optional(),
+  userPropertiesSupportAgentName: z.nullable(z.string()).optional(),
 }).transform((v) => {
   return remap$(v, {
     idHour: "id_hour",